Fix +100% buff issue, changed multiplier type to float, added more raid sizes
Buffs were getting an additional +100% due to HandleStatModifier() treating percentages as delta rather than whole.  As a result, a 2x multiplier was applying a +300% buff.
Changing multiplier type to float allows more fine tuning, ie. allowing 1.5 if 1 is too hard and 2 is too easy.
Added support for Mythic 5 mans and 15 and 30 man raids.