DK Starter Area Fix. Play Thru Audit
There is a few times i had to relog so that phasing to be updated. Keep that in mind.
Fixed Fineous NPC. Clicking on it gave no speech gossip. Now when you click on him you get the sell menu and the repair options.
Quest start NPC for Quest 12779 was not spawned anywhere, phasing was screwy but a different lich king id was present but was falling thru the ramp. so i reposition him at the base while making him the new quest starter so progression does not break.