DK Portal Fix pt2 and Swirl Racial Swap NPC
DK Portals are now proper fixed at the end of the DK Starter zone quest line.
Racial Trait Swapper NPC "SWIRL", my own personal NPC from 335 TC Custom branch is now incorporated and upgraded and modified to reflect the Racial swaps of Cata and to include the Worgen and Goblin Races and new classes to existing races. Conf is adjustable to change up the cost in gold for the racial swap.