Bandaid Workaround script started
This is the bandaid script
Designed to just assist with progress for whatever scenario\quest line that is need.
This script is a hit subject to change. If sql is include please make notes of it.
Currently only have two gilneas quests for now. Quest 36440 and 36540, I am not famliar with sql so these temporary scripts will work for now. still not done auditing worgen starterzone.