Gilneas is Completable now with no GM Cmds
Remember phasing is a issue so relog if things are suspect and not displaying.
A few missions had to be set to auto complete due to missing core data, and a couple are disabled due to missing core data but can't be auto completed.
Character scripted made for Battle of gilneas.