Aoe Loot and Mail Excess Conf
New QOL and Conf update:
AOE.LOOT.enable
If player is not in a group, enables group looting of 30 and mails excess to player.
If too much is mailed, it will not display all messages so clear your inbox and relog to acquired a refresh listed of items that are queued.